try to fix the rounding errors in the database

+++ backend/src/main/java/com/github/ulrikewerner/backend/entities/Card.java
@@ -3,6 +3,6 @@ package com.github.ulrikewerner.backend.entities;
 public record Card(
         String id,
         String team,
-        double pointsPerGame
+        int pointsPerGame
 ) {
 }
